[KAIZEN-0] splittet sagas i egne filer
